grep: remove check_multibyte_string, fix non-UTF8 missed match
Avoid computing ahead something that can be computed lazily as efficiently
(or more efficiently in the case of UTF-8, though this is left as TODO).
At the same time, "soften" the rejection condition for matching in the
middle of a multibyte sequence to fix bug 23814.

Multibyte "grep -i" would still be very slow if it wasn't for the workaround
patch c32c042 (grep: match multibyte charsets line-by-line when using -i,
2010-03-08).

* NEWS: Document bugfix.
* src/search.c (check_multibyte_string): Rewrite as...
(is_mb_middle): ... this.
(EGexecute, Fexecute): Adjust.
* tests/Makefile.am (TESTS): Add euc-mb.
* tests/euc-mb: New testcase.