grep -F: fix a multi-byte erroneous-match-in-middle bug
Just as Perl prints nothing in this case,
  printf '\357\274\241\n' | perl -CIO -lne '/\357/ and print'

grep should also print nothing when used as follows.
However, these would mistakenly match with grep prior to 2.6.2:
  printf '\357\274\241\n' | LC_ALL=en_US.UTF-8 src/grep -F $'\357'
  printf '\357\274\241\n' | LC_ALL=en_US.UTF-8 src/grep -F $'\357\274'

* src/searchutils.c (is_mb_middle): New parameter: the length of the
match, in bytes, as determined by kwsexec.  Use this to detect when
the nominal match found by kwsexec must be skipped because it is for
an incomplete multi-byte character that is a prefix of a character
in the input.
* src/dfasearch.c (EGexecute): Update caller.
* src/kwsearch.c (Fexecute): Likewise.
* src/search.h: Update prototype.
* NEWS (Bug fixes): Mention it.
Report and analysis by Norihiro Tanaka.