dfa: fall back to glibc matcher if a MBCSET is found
This patch enables full support of equivalence classes and multicharacter
collation symbols.  It can also improve performance problems in some
cases for multibyte grep.  Both of these changes however depend on the
glibc version installed in the system.

For UTF-8 it will trigger only in the presence of MBCSET, e.g. [a-z].
For other character sets all brackets and `.` as well will trigger it.

* NEWS: Document this.
* src/dfa.c (dfaexec): Fall back to glibc for multibyte matches,
if possible.