dfa: process range expressions consistently with system regex
The actual meaning of range expressions in glibc is not exactly strcoll,
which makes the behavior of grep hard to predict when compiled with the
system regex.  Leave to the system regex matcher the decision of which
single-byte characters are matched by a range expression.

This partially reverts a change made in commit 0d38a8bb (which made
sense at the time, but not now that src/dfa.c is not doing multibyte
character set matching anymore).

* src/dfa.c (in_coll_range): Remove.
(parse_bracket_exp): Use system regex to find which single-char
bytes match a range expression.