tests: support development version of automake too
This change implements a more correct and idiomatic use of the
features of the Automake-provided 'parallel-tests' harness.
Moreover, this change is required in order for the testsuite to
continue to work with the new testsuite harness that is planned
to be introduced in Automake 1.12 (which, as of the writing date,
is still under development and in late alpha state).

* tests/Makefile.am (TESTS_ENVIRONMENT): The development version of
automake dos not support setting the interpreter delegated to run
the tests scripts in this variable; instead, use ...
(LOG_COMPILER): ... this variable.
* .gitignore: Ignore `.trs' files in directory `tests/'.
* build-aux/.gitignore: Ignore `test-driver' script.