grep: do input==output check more like dir loop check
* src/main.c (grepfile): Just use SAME_INODE; don't bother
with SAME_REGULAR_FILE.  This works better on properly-working
POSIX hosts, since it handles the case where the file is changing
as we grep it.  It works worse on hosts that don't support st_ino
properly, but in practice this isn't that much of a problem here.
* src/system.h (same_file_attributes, SAME_REGULAR_FILE):
Remove; no longer needed.