fgrep: fix case-fold incompatibility with plain 'grep'
fgrep converted to lowercase, whereas the regex code converted
to uppercase.  The resulting behaviors don't agree in offbeat
cases like Greek sigmas and Turkish Is.  Fix this by changing
fgrep to agree with the regex code.
* src/kwsearch.c (Fcompile, Fexecute):
* src/searchutils.c (kwsinit, mbtoupper):
Convert to uppercase, not to lowercase, for compatibility with
plain 'grep'.
* src/search.h, src/searchutils.c (mbtoupper):
Rename from mbtolower, since it now converts to uppercase.
All uses changed.
* tests/case-fold-titlecase: Add tests for this.