grep -F -x -o PAT would print an extra newline for each match
* src/kwsearch.c (Fexecute): Correctly compute the length of a match
by subtracting 2 (not 1) when match_lines is set.  With -x, we augment
the "line" by both prepending and appending an EOLBYTE to the search
pattern.  Here, we must correct for that.  However, to compensate,
when we are using -x (--line-regexp) and start_ptr is NULL, we have
to add 1 to the length so that we still print the trailing EOLBYTE.
Introduced by commit v2.18-85-g2c94326.
* tests/match-lines: Add a new test.
* tests/Makefile.am (TESTS): Add it.
* NEWS (Bug fixes): Mention it.