dfa: speed up handling of long pattern
DFA tries to find a long sequence of characters that must appear
in any matching line.  However, when a pattern is long (length N),
it is very slow, because it makes O(N^2) strstr calls.
This change reduces that to O(N) by processing each sequence of
adjacent "regular" characters as a group.

Compare the run times of this command before and after this change:
(on a i7-4770S CPU @ 3.10GHz using rawhide (~fedora 22) and compiled
with gcc 6.0.0 20150627)
  : | env time -f %e grep -f <(seq -s '' 9999)
  Before: 0.85
   After: 0.02

* src/dfa.c (dfamust): Process each string of concatenated normal
characters as a unit.
* NEWS (Improvement): Mention it.
Prompted by a bug report and patch by Ivan Yanikov
in http://bugs.gnu.org/15191#5